requirements.txt
playwright==1.49.1
pandas==2.2.3
PySide6==6.8.1.1
google-generativeai==0.8.4
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
beautifulsoup4==4.12.3
requests==2.32.3
dnspython==2.7.0
fastapi==0.116.1
uvicorn==0.35.0
